Centralize navigators code.
