+++ src/navigation/navigators/all.ts
@@ -1,0 +1,4 @@
+import { stackBuilder } from './stack'
+import { switchBuilder } from './switch'
+
+export const navigatorBuilders = [stackBuilder, switchBuilder] as const
