+++ src/navigation/types.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react'
+import { Navigator } from './navigators'
 
 /**
  * UnionToIntersection
@@ -15,40 +16,7 @@ type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (
   ? I
   : never
 
-export interface StackNavigatorOptions {
-  animation?: 'none' | 'default'
-}
-
-export interface StackNavigator {
-  readonly stack: NavigatorsOrScreens
-  readonly options?: StackNavigatorOptions
-}
-
-export interface SwitchNavigator {
-  readonly switch: NavigatorsOrScreens
-}
-
-export type Navigator = StackNavigator | SwitchNavigator
-type NavigatorsOrScreens = readonly (Screen<string> | Navigator)[]
-export type Screen<Name extends string> = {
-  [P in Name]: React.ComponentType<any>
-}
-
 export type NavigationSchema = Navigator
-
-type Screens<T extends NavigatorsOrScreens> = {
-  [P in keyof T]: T[P] extends Screen<string>
-    ? T[P]
-    : T[P] extends Navigator
-    ? NavigatorScreens<T[P]>
-    : never
-}[Extract<keyof T, number>]
-
-export type NavigatorScreens<T extends Navigator> = T extends StackNavigator
-  ? Screens<T['stack']>
-  : T extends SwitchNavigator
-  ? Screens<T['switch']>
-  : never
 
 export type ScreenArguments<T> = {
   [P in Extract<keyof UnionToIntersection<T>, string>]: UnionToIntersection<
